+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed 
+- Added the `AxiosStatic` object as a `Bridge` static property to avoid potential faults because of imports.
 
 ## 1.0.8 - 2022-11-11
 ### Changed
@@ -14,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 1.0.7 - 2022-11-11
 ### Added
-- Added `connection_timeout` as a `Bridge` instance
+- Added `connection_timeout` as a `Bridge` property
 - Generated documentation
 - Updated default exported axios item access to check for the item on the default instance